// typechain-types/PolicyRegistry.ts
export declare namespace PolicyRegistry {
  export type PolicyStruct = {
    id: BytesLike;
    name: string;
    maxAutoAmount: BigNumberish;
    requireManualReviewAbove: BigNumberish;
    requireWhitelist: boolean;
    active: boolean;
    owner: AddressLike;
    createdAt: BigNumberish;
    updatedAt: BigNumberish;
  };

  export type PolicyStructOutput = [
    id: string,
    name: string,
    maxAutoAmount: bigint,
    requireManualReviewAbove: bigint,
    requireWhitelist: boolean,
    active: boolean,
    owner: string,
    createdAt: bigint,
    updatedAt: bigint
  ] & {
    id: string;
    name: string;
    maxAutoAmount: bigint;
    requireManualReviewAbove: bigint;
    requireWhitelist: boolean;
    active: boolean;
    owner: string;
    createdAt: bigint;
    updatedAt: bigint;
  };
}